+++ src/app/(withoutNavBar)/product-list/product/[id]/page.tsx
@@ -1,9 +1,7 @@
 import Header from "@/components/productDetail/Header";
-import Description from "@/components/productDetail/Description";
 import { Suspense } from "react";
 import Skeleton from "@/components/productDetail/Skeleton";
-import DetailPrice from "@/components/productDetail/DetailPrice";
-import RelatedProducts from "@/components/productDetail/RelatedProducts";
+import DetailDataFetcher from "@/components/productDetail/DetailDataFetcher";
 
 type Props = {
   params: {
@@ -16,9 +14,7 @@ export default function ProductDetailPage({ params: { id } }: Props) {
     <section className="flex flex-col h-full overflow-y-auto">
       <Header />
       <Suspense fallback={<Skeleton />}>
-        <Description id={id} />
-        <DetailPrice id={id} />
-        <RelatedProducts id={id} />
+        <DetailDataFetcher id={id} />
       </Suspense>
     </section>
   );
