+++ src/components/productDetail/Description.tsx
@@ -1,18 +1,8 @@
-"use client";
-
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { getProductFetcher } from "@/fetchers/product";
 import ProductImage from "../products/wishList/ProductImage";
 import Information from "./Information";
+import { DetailProduct } from "@/models/product";
 
-export default function Description({ id }: { id: string }) {
-  const productId = Number(id);
-  // productId가 number 타입인지 확인 후 not-found로 리다이렉트
-  const { data: product } = useSuspenseQuery({
-    queryKey: ["product", productId],
-    queryFn: () => getProductFetcher(productId),
-  });
-
+export default function Description({ product }: { product: DetailProduct }) {
   return (
     <div className="p-4">
       <ProductImage
