+++ src/components/productDetail/RelatedProducts.tsx
@@ -1,17 +1,8 @@
-"use client";
-
-import { getRelatedProductsFetcher } from "@/fetchers/product";
-import { useSuspenseQuery } from "@tanstack/react-query";
 import RelatedProductItem from "./RelatedProductItem";
 import Carousel from "../common/Carousel";
+import { Product } from "@/models/product";
 
-export default function RelatedProducts({ id }: { id: string }) {
-  const productId = Number(id);
-  const { data: products } = useSuspenseQuery({
-    queryKey: ["product", productId, "related_product_list"],
-    queryFn: () => getRelatedProductsFetcher(productId),
-  });
-
+export default function RelatedProducts({ products }: { products: Product[] }) {
   return (
     <div className="flex flex-col gap-3 p-4 border-t-4 border-border">
       <h3 className="font-bold text-black leading-7">
